requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy>=1.26.1
Pillow
huggingface_hub
einops
safetensors
scipy
termcolor
matplotlib
tqdm
requests
